fixed api and dal connection, still mapping error

+++ dms-api/Models/Document.cs
@@ -2,9 +2,9 @@
 {
     public class Document
     {
-        public long Id {  get; set; }
+        public int Id {  get; set; }
 
-        public Document(long id)
+        public Document(int id)
         {
             Id = id;
         }
